Controllers for All Cars and by Id's

+++ src/main/java/com/project/autoServiceApp/repository/CarRepository.java
@@ -3,5 +3,6 @@ package com.project.autoServiceApp.repository;
 import com.project.autoServiceApp.model.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
+
 }
